# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API для платформы по усыновлению домашних животных.",
        "title": "Happytail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/animals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Получить всех животных",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Animal"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Создать животное",
                "parameters": [
                    {
                        "description": "данные животного",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.CreateAnimalInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Animal"
                        }
                    }
                }
            }
        },
        "/animals/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Получить животное по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID животного",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Animal"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update existing animal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Update animal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update animal input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.UpdateAnimalInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Animal"
                        }
                    },
                    "400": {
                        "description": "Invalid input or invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Animal not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет животное по ID (soft delete)",
                "tags": [
                    "animals"
                ],
                "summary": "Удалить животное",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID животного",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно удалено"
                    },
                    "400": {
                        "description": "invalid animal ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Animal not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete animal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/posts": {
            "get": {
                "description": "Возвращает список всех постов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Получить все посты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Post"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch posts",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создаёт новый пост",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Создать пост",
                "parameters": [
                    {
                        "description": "данные поста",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.CreatePostInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create post",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "description": "Возвращает один пост по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Получить пост по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID поста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Post"
                        }
                    },
                    "400": {
                        "description": "Invalid post ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch post",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/posts/{id}/status": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Обновить статус поста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID поста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "новый статус",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.UpdatePostStatusInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_sd0hni-psina_happytail_internal_models.Animal": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_vet_passport": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_vaccinated": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "share_count": {
                    "type": "integer"
                },
                "shelter_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.AnimalType"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.AnimalType": {
            "type": "string",
            "enum": [
                "dog",
                "cat",
                "bunny"
            ],
            "x-enum-varnames": [
                "AnimalTypeDog",
                "AnimalTypeCat",
                "AnimalTypeBunny"
            ]
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.CreateAnimalInput": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "animal_type": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.AnimalType"
                },
                "breed": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_vet_passport": {
                    "type": "boolean"
                },
                "is_vaccinated": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "shelter_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_sd0hni-psina_happytail_internal_models.CreatePostInput": {
            "type": "object",
            "properties": {
                "animal_id": {
                    "type": "integer"
                },
                "contact_info": {
                    "type": "string"
                },
                "listing_type": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.ListingType"
                },
                "photo_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Money"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.ListingType": {
            "type": "string",
            "enum": [
                "sale",
                "give"
            ],
            "x-enum-varnames": [
                "ListingTypeSale",
                "ListingTypeGive"
            ]
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.Money": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.Post": {
            "type": "object",
            "properties": {
                "animal_id": {
                    "type": "integer"
                },
                "contact_info": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "listing_type": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.ListingType"
                },
                "photo_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.Money"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.PostStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_sd0hni-psina_happytail_internal_models.PostStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "deleted"
            ],
            "x-enum-varnames": [
                "PostStatusActive",
                "PostStatusInactive",
                "PostStatusDeleted"
            ]
        },
...
        "github_com_sd0hni-psina_happytail_internal_models.UpdateAnimalInput": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "breed": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_vet_passport": {
                    "type": "boolean"
                },
                "is_vaccinated": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_sd0hni-psina_happytail_internal_models.UpdatePostStatusInput": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/github_com_sd0hni-psina_happytail_internal_models.PostStatus"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}